Run keepalive test exclusively on Linux
- keepalive tests are specified at compile time, not runtime

fixes (Windows)
```
httpclient\httpclient\default_http_clients_test.go:54: cannot use int(f.Fd()) (type int) as type syscall.Handle in argument to syscall.GetsockoptInt
httpclient\httpclient\default_http_clients_test.go:57: undefined: keepaliveGetsockoptPlatformIndependent
httpclient\httpclient\default_http_clients_test.go:59: undefined: keepaliveIntervalPlatformIndependent
httpclient\httpclient\default_http_clients_test.go:59: cannot use int(f.Fd()) (type int) as type syscall.Handle in argument to syscall.GetsockoptInt
```

[#143754851](https://www.pivotaltracker.com/story/show/143754851)
