+++ httpclient/default_http_clients_test.go
@@ -8,61 +8,11 @@ import (
 
 	. "github.com/cloudfoundry/bosh-utils/httpclient"
 	"github.com/cloudfoundry/bosh-utils/httpclient/fakes"
-	"syscall"
-	"net"
-	"os"
-	"runtime"
 )
 
 var _ HTTPClient = &fakes.FakeHTTPClient{}
 
 var _ = Describe("Default HTTP clients", func() {
-	if runtime.GOOS == "linux" || runtime.GOOS == "darwin" {
-		It("enables TCP (socket) keepalive with an appropriate interval", func() {
-			// to test keepalive, we need a socket. A socket is an _active_ TCP connection to a server.
-			// we make our own server, connect to it, and make our assertions against the socket
-			laddr := "127.0.0.1:19642" // unlikely-to-be-used port number, unprivileged (1964, Feb, my birth)
-			readyToAccept := make(chan bool, 1)
-
-			go func() {
-				defer GinkgoRecover()
-				defer func(){
-					readyToAccept<-true
-				}()
-
-				ln, err := net.Listen("tcp", laddr)
-				Expect(err).ToNot(HaveOccurred())
-
-				readyToAccept<-true
-
-				_, err = ln.Accept()
-				Expect(err).ToNot(HaveOccurred())
-			}()
-
-			<-readyToAccept
-
-			client := CreateDefaultClient(nil)
-			connection, err := client.Transport.(*http.Transport).Dial("tcp", laddr)
-			Expect(err).ToNot(HaveOccurred())
-
-			tcpConn, ok := connection.(*net.TCPConn)
-			Expect(ok).To(BeTrue())
-
-			f, err := tcpConn.File()
-			Expect(err).ToNot(HaveOccurred())
-
-			sockoptValue, err := syscall.GetsockoptInt(int(f.Fd()), syscall.SOL_SOCKET, syscall.SO_KEEPALIVE)
-			err = os.NewSyscallError("getsockopt", err)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(sockoptValue).To(Equal(keepaliveGetsockoptPlatformIndependent))
-
-			sockoptValue, err = syscall.GetsockoptInt(int(f.Fd()), syscall.IPPROTO_TCP, keepaliveIntervalPlatformIndependent)
-			err = os.NewSyscallError("getsockopt", err)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(sockoptValue).To(Equal(30))
-		})
-	}
-
 	Describe("DefaultClient", func() {
 		It("is a singleton http client", func() {
 			client := DefaultClient
