--- httpclient/keepalive_syscall_darwin_test.go
@@ -1,6 +1,0 @@
-package httpclient_test
-
-import "syscall"
-
-const keepaliveIntervalPlatformIndependent = syscall.TCP_KEEPALIVE
-const keepaliveGetsockoptPlatformIndependent = syscall.SO_KEEPALIVE
